uric_api/requirements.txt
asgiref==3.5.2
bcrypt==4.0.1
cffi==1.15.1
cryptography==38.0.3
Django==3.2.16
django-cors-headers==3.13.0
django-simpleui==2022.7.29
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
et-xmlfile==1.1.0
openpyxl==3.0.10
paramiko==2.12.0
Pillow==9.3.0
pycparser==2.21
PyJWT==1.7.1
PyMySQL==1.0.2
PyNaCl==1.5.0
pytz==2022.6
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.4.0
